+++ petshop/src/main/java/com/ccdjmv/petshop/repository/CartRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ccdjmv.petshop.entity.CartEntity;
 
 @Repository
-public interface CartRepository extends JpaRepository<CartEntity, Integer>{
+public interface CartRepository extends JpaRepository<CartEntity, Long>{
 	
 }
